Refactors reputation to use floating point
Changes the data type of `reputation` fields from integers to floating-point numbers (f64) in multiple structs related to user data.

This change allows for more granular representation of user reputation scores.

+++ src/types/signin_response.rs
@@ -10,7 +10,7 @@ pub struct SigninResponse {
     pub background: Option<String>,
     pub crossplay: bool,
     pub platform: String,
-    pub reputation: i64,
+    pub reputation: f64,
     pub has_mail: bool,
     pub region: String,
     pub written_reviews: i64,
