+++ src/types/user_short.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::enums::*;
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
 pub struct UserShort {
     pub id: String,
     #[serde(rename = "ingame_name", alias = "ingameName")]
@@ -10,7 +10,7 @@ pub struct UserShort {
     /// Optional link to the user's avatar image.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub avatar: Option<String>,
-    pub reputation: i32,
+    pub reputation: f64,
     #[serde(rename = "status", default = "StatusType::default")]
     pub status: StatusType,
 }
